requirements.txt
streamlit
pandas
spacy
scikit-learn
wordcloud
networkx
python-louvain
openpyxl
python-pptx
matplotlib